refactor : 기능 코드 리팩토링
