--- src/main/java/nextstep/subway/domain/PathFinder.java
@@ -1,6 +1,0 @@
-package nextstep.subway.domain;
-
-public interface PathFinder {
-
-  Path findPath(Station source, Station target);
-}
