+++ src/test/java/nextstep/subway/unit/LineServiceTest.java
@@ -31,7 +31,7 @@ class LineServiceTest {
         Station 삼성역 = stationRepository.save(new Station("삼성역"));
         Line 이호선 = lineRepository.save(createLine(강남역, 역삼역));
 
-        lineService.addSection(이호선.getId(), new SectionRequest(역삼역.getId(), 삼성역.getId(), 10));
+        lineService.addSection(이호선.getId(), new SectionRequest(역삼역.getId(), 삼성역.getId(), 10, 10));
 
         Line line = lineService.findById(이호선.getId());
