+++ src/test/java/nextstep/subway/unit/PathServiceMockTest.java
@@ -10,7 +10,7 @@ import nextstep.subway.applicaion.PathService;
 import nextstep.subway.applicaion.StationService;
 import nextstep.subway.applicaion.dto.PathResponse;
 import nextstep.subway.domain.Line;
-import nextstep.subway.domain.PathType;
+import nextstep.subway.domain.path.PathType;
 import nextstep.subway.domain.Station;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
